fix: moved InfoBadge to common; adjusted About section header placement to center on all screens
