+++ src/common/InfoBadge.tsx
