+++ src/common/index.ts
@@ -2,4 +2,5 @@ export * from "./ContactButton"
 export * from "./EmailButton"
 export * from "./FullScreenDialog"
 export * from "./GitHubButton"
+export * from "./InfoBadge"
 export * from "./LinkedInButton"
