+++ src/features/about/index.ts
@@ -1,2 +1,1 @@
 export * from "./About"
-export * from "./InfoBadge"
